testing docker compose up -d test

+++ docker_first_deploy.sh
@@ -78,10 +78,12 @@ DOCKER_SETTINGS_NEW="$(
 )"
 
 echo "Warning: This script will stop all running containers and then deploy nftables rules for the built application.  It will also generate an archive of logs from the execution and containers for debugging purposes."
-echo "${HOSTNAME}"
-read -erp 'Continue? ' -i 'yes' continue
-if ! [[ ${continue} == 'yes' ]]; then
-   help7
+if [[ "${GITHUB_ACTIONS:-}" == "true" ]]; then
+   echo "Běžíme v GitHub Actions"; else
+   read -erp 'Continue? ' -i 'yes' continue
+   if ! [[ ${continue} == 'yes' ]]; then
+      help7
+   fi
 fi
 
 rm -rf "${LOG_DIR}" "${NFTABLES}"
